chore: refactor content styling
