+++ packages/gui/src/components/Dashboard/TabRecordings.vue
