+++ packages/gui/src/js/router.js
@@ -1,5 +1,5 @@
+import Dashboard from '../components/Dashboard/Dashboard.vue'
 import Login from '../components/Login.vue'
-import Manager from '../components/Manager/Manager.vue'
 import Settings from '../components/Settings/Settings.vue'
 import Splash from '../components/Splash.vue'
 import StreamView from '../components/StreamView.vue'
@@ -8,9 +8,9 @@ import {createRouter, createWebHistory} from 'vue-router'
 export default function(app) {
     const routes = [
         {
-            component: Manager,
-            name: 'manager',
-            path: '/manager/:tabId',
+            component: Dashboard,
+            name: 'dashboard',
+            path: '/dashboard/:tabId',
         },
         {
             component: Settings,
