+++ packages/gui/src/scss/elements/_buttons.scss
@@ -9,7 +9,11 @@
     transition: all ease 0.5s;
 
     &[disabled] {
-        color: var(--grey-5);
+        color: var(--grey-6);
+    }
+
+    &.active {
+        color: var(--primary-color);
     }
 
     &:focus {
@@ -44,7 +48,6 @@
         }
 
         &[disabled] {
-            color: var(--grey-4);
 
             &:hover {
                 cursor: not-allowed;
@@ -82,6 +85,7 @@
         }
 
         &.active {
+            border-bottom: var(--border) solid var(--primary-color);
             color: var(--primary-color);
         }
     }
