+++ packages/gui/src/scss/elements/_presence.scss
@@ -10,18 +10,23 @@
         font-weight: bold;
         height: var(--space-4);
         margin-bottom: var(--spacer);
+        padding-right: var(--spacer);
         position: relative;
 
-        svg {
-            color: var(--primary-color);
-            margin-right: var(--spacer);
+        .logo {
+
+            svg {
+                color: var(--primary-color);
+                margin-right: var(--spacer);
+            }
+
         }
 
         .version {
             bottom: 0;
             font-size: var(--text-small);
             position: absolute;
-            right: var(--spacer);
+            right: calc(var(--spacer) * 1.75);
         }
     }
